--- tests/unit/backend/recipe/example.cif
@@ -1,32 +1,0 @@
-# mock CIF file based on Silicon_NIST_640f.cif in analysis cluster
-data_mock
-_chemical_name_common 'Si'
-_chemical_formula_structural 'Si'
-_cell_length_a 5.43159(20)
-_cell_length_b 5.43159(20)
-_cell_length_c 5.43159(20)
-_cell_angle_alpha 90.
-_cell_angle_beta 90.
-_cell_angle_gamma 90.
-_cell_volume 160.24
-_cell_formula_units_Z 8
-_space_group_name_H-M_alt 'F d -3 m S'
-_space_group_IT_number 227
-loop_
-_atom_type_symbol
-_atom_type_oxidation_number
-La0+ 0
-B0+ 0
-loop_
-_atom_site_label
-_atom_site_type_symbol
-_atom_site_symmetry_multiplicity
-_atom_site_Wyckoff_symbol
-_atom_site_fract_x
-_atom_site_fract_y
-_atom_site_fract_z
-_atom_site_U_iso_or_equiv
-_atom_site_occupancy
-La1 La0+ 1 a 0 0 0 0.0016(2) 1.
-B1 B0+ 6 f 0.2018(4) 0.5 0.5 0.0027(3) 1.
-#End of TTdata_120585-ICSD
